Added points count for a deck

+++ src/main/java/fr/arolla/card/Card.java
@@ -6,5 +6,5 @@ public abstract class Card {
 
     public abstract boolean hasAHigherRankThat(Card card);
 
-    public abstract int getPoints();
+    public abstract double getPoints();
 }
